feature add team_member schemas & computed field & assign manager/ TOFIX add_employee function to add to team_member controller(TOCREATE)

+++ time_manager/priv/repo/migrations/20241025134116_add_hours_per_week.exs
@@ -3,7 +3,7 @@ defmodule TimeManager.Repo.Migrations.AddHoursPerWeek do
 
   def change do
     alter table(:users) do
-      add :hours_per_day, :integer, default: 8
+      add :hours_per_day, :float, default: 8.0
     end
   end
 end
